requirements.txt
openai-whisper
torch
numpy
tqdm
more-itertools
numba
ffmpeg-python